Update IPC.cs
I changed my mind on how to handle versions and provide two API Versions now, the major one will only change on breaking changes, while the minor changes on any addition, too.
So for your use case we check for equality on the major again. 
This change is not breaking since it's a new IPC function, so this should remain working even before you accept the PR, but give a log warning when the old Penumbra.ApiVersion is called.
So you don't need to hurry or anything. :)